+++ python/sensors/Travel.py
@@ -1,5 +1,5 @@
 # from MPU6050 import mpu6050
-from MPU6050_improved import mpu6050_safe as mpu6050
+from lib_movit_sensors.MPU6050_improved import mpu6050_safe as mpu6050
 from lib_movit_sensors.IMUDetectMotion import IMUDetectMotion
 from datetime import datetime
 import asyncio
